+++ changelog.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Features
+
 - Add `pkg/cosmostxcollector` package with support to query and save TXs and events.
 - Add `ignite network coordinator` command set.
 - Add `ignite network validator` command set.
@@ -26,6 +27,7 @@
 - Add `--peer-address` flag to `network chain join` command.
 - Change nightly tag format
 - Add cosmos-sdk version in `version` command
+- [#29350](https://github.com/ignite/cli/pull/2935) Update `gobuffalo/plush` templating tool to `v4`
 
 ### Fixes